test(server): split handlers_test.go to match the per-flow source files
handlers_test.go tested code that now lives across many files. Split it into
test files matching the source layout (discovery_test, authorize_test,
login_test, approval_test, connectorauth_test, grant_*_test) with shared
fixtures in helpers_test.go, folding handlers_approval_test.go into
approval_test.go and health tests into server_test.go. Pure test movement:
same 176 tests, all passing.
